requirements.txt
Django>=4.2,<5.0
Pillow>=10.0.0
djangorestframework>=3.14.0
django-filter>=23.0
django-cors-headers>=4.0.0
python-decouple>=3.8
gunicorn>=21.0.0
psycopg2-binary>=2.9.0
celery>=5.3.0
redis>=5.0.0
stripe>=7.0.0
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0

